refactor(lsp): refactor
